Fix #161897. Better init height estimate.
